use current time oid as default
If we tried to run this bot on a new machine with an old account, right
now it would go back and process all of the old actions again. We want
to avoid that, so we'll cut our losses and just use the current time as
the starting tail marker.